blockstack_resolver/requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
commontools==0.1.0
itsdangerous==0.24
lockfile==0.10.2
pylibmc==1.3.0
python-bitcoinrpc==0.1
pybitcoin==0.8.2
pymongo==3.0.2
requests==2.3.0
beautifulsoup4==4.3.2